% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe_phonemes.R
\docType{data}
\name{english_phonemes}
\alias{english_phonemes}
\title{English phoneme inventory (IPA)}
\format{
An object of class \code{character} of length 39.
}
\usage{
english_phonemes
}
\description{
The 39 phonemes of the CMUdict English inventory, in the IPA forms produced
by \code{build_cmu_ipa()}. Pass this as \code{inventory} in \code{code_phonemes()} or
\code{transcribe_phonemes()} so that separately coded datasets end up with
identical columns.
}
\keyword{datasets}
